requirements.txt
PyYAML
discord-py-interactions
lavalink